Remove route usernames
As this fragment was redundant anyway due to the middleware, it is
now removed in favour of checking the jwt username from the request.

Body parameters for post requests have been restructured to be
explicitly required in the route file to remove the need for useless
body wrapping on the frontend.